requirements.txt
asgiref==3.8.1
Django==4.2.14
gunicorn==22.0.0
mssql-django==1.5
mysqlclient==2.2.4
packaging==24.1
pyodbc==5.1.0
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.5.1
tzdata==2024.1
whitenoise==6.7.0
